Modifying some models

+++ Models/OrderDetail.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
+using System.ComponentModel.DataAnnotations.Schema;
 using System.Linq;
 using System.Threading.Tasks;
 
@@ -14,7 +15,8 @@ namespace RestaurantApi.Models
        
         public int FoodItemId { get; set; }
         public FoodItem FoodItem { get;set; }
-        public decimal FoodItemPrice { get; set; }
+
+        public int FoodItemPrice { get; set; }
         public int Quantity { get; set; }
 
     }
